Ajustando arquivos apontados no deply do render

+++ src/expense/expense.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Expense } from './expense.entity';
@@ -10,32 +10,38 @@ export class ExpenseService {
     private expenseRepository: Repository<Expense>,
   ) {}
 
-  findAll(): Promise<Expense[]> {
-    return this.expenseRepository.find();
+  async findAll(userId: number): Promise<Expense[]> {
+    return this.expenseRepository.find({ where: { user: { id: userId } } });
   }
-
-  create(expense: Partial<Expense>): Promise<Expense> {
-    const newExpense = this.expenseRepository.create(expense);
+  
+  async create(expense: Partial<Expense>, userId: number): Promise<Expense> {
+    const newExpense = this.expenseRepository.create({ ...expense, user: { id: userId } });
     return this.expenseRepository.save(newExpense);
   }
 
-  // Método para consultar por intervalo de data (usando QueryBuilder)
-  async findByDateRange(startDate: string, endDate: string): Promise<Expense[]> {
+  async findByDateRange(startDate: string, endDate: string, userId: number): Promise<Expense[]> {
     return this.expenseRepository
       .createQueryBuilder('expense')
       .where('expense.date >= :startDate', { startDate })
       .andWhere('expense.date <= :endDate', { endDate })
+      .andWhere('expense.user.id = :userId', { userId })
       .getMany();
   }
 
-  // Atualizar um registro
-  async update(id: number, expense: Partial<Expense>): Promise<Expense> {
-    await this.expenseRepository.update(id, expense);
-    return this.expenseRepository.findOne({ where: { id } });
+  async update(id: number, expense: Partial<Expense>, userId: number): Promise<Expense> {
+    const existingExpense = await this.expenseRepository.findOne({ where: { id, user: { id: userId } } });
+    if (!existingExpense) {
+      throw new NotFoundException('Despesa não encontrada ou não pertence ao usuário.');
+    }
+    Object.assign(existingExpense, expense);
+    return this.expenseRepository.save(existingExpense);
   }
 
-  // Excluir um registro
-  async delete(id: string): Promise<void> {
+  async delete(id: string, userId: number): Promise<void> {
+    const expense = await this.expenseRepository.findOne({ where: { id: +id, user: { id: userId } } });
+    if (!expense) {
+      throw new NotFoundException('Despesa não encontrada ou não pertence ao usuário.');
+    }
     await this.expenseRepository.delete(id);
   }
 }
